msm-adreno-tz: Introduce adrenoboost v3.0
Credits to flar2 for the original work, and tbalden's
for the boosting tweaks and improvement.

This commit updates the device frequency scaling algorithm by adding
three different level maps and adrenoboost parameter. The level maps
contain multiplicator and divider values for each level that affect
the busy time calculation. The adrenoboost parameter adjusts the busy
time calculation by adding a multiplier to the formula. Additionally,
the code includes logic for level conservation and adjustment based on
the input value. The commit also resets the device frequency scaling
parameters after level adjustment.

Please note that the GPU boosting mechanism in this code is "hacky" and
placed under ifdefs and is not advisable for better GPU models. Boosting
the GPU can significantly impact battery life, despite the presence of
power-saving measures within the code.

Signed-off-by: mcdofrenchfreis <xyzevan@androidist.net>